Refactor: move business logic out of reader
