+++ xnb/objects/__init__.py
@@ -1,0 +1,2 @@
+
+from .Texture2D import Texture2D
